Works now from a single frontend file and also unified some of the backend
