+++ src/fish-catch-map/block.json
@@ -67,6 +67,5 @@
     "textdomain": "fish-catch",
     "editorScript": "file:./index.js",
     "editorStyle": "file:./index.css",
-    "style": "file:./style-index.css",
-    "viewScript": "file:./frontend.js"
+    "style": "file:./style-index.css"
 }
